+++ backend/cmd/main.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	_ "github.com/lib/pq"
+	"github.com/yoshihiro-shu/draft-backend/backend/infrastructure/cache"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/cache"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/model"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/auth"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/config"
