+++ backend/interfaces/api/handler/top_page.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/yoshihiro-shu/draft-backend/backend/application/usecase"
 	"github.com/yoshihiro-shu/draft-backend/backend/domain/model"
+	"github.com/yoshihiro-shu/draft-backend/backend/infrastructure/cache"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/request"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/cache"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/logger"
 	"go.uber.org/zap"
 )
