+++ backend/interfaces/api/server/server.go
@@ -8,9 +8,9 @@ import (
 	"os/signal"
 	"time"
 
+	"github.com/yoshihiro-shu/draft-backend/backend/infrastructure/cache"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/router"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/user_api"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/cache"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/model"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/config"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/logger"
