+++ backend/interfaces/api/user_api/api.go
@@ -1,11 +1,11 @@
 package user_api
 
 import (
+	"github.com/yoshihiro-shu/draft-backend/backend/infrastructure/cache"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/handler"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/middlewares"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/request"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/router"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/cache"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/model"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/config"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/logger"
