+++ backend/internal/mock_test/mock_cache.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/alicebob/miniredis/v2"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/cache"
+	"github.com/yoshihiro-shu/draft-backend/backend/infrastructure/cache"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/config"
 )
